backend/requirements.txt
docx2txt
wikipedia
pdfplumber
#wikipediaapi==0.5.8
langchain
scikit-learn
pathlib
elasticsearch
nltk
python-docx
torch
transformers
sentencepiece
altair==4.2.2
flask
flask-cors